Initializes modules
This gets the DOM module in okay shape.

To test DOM related things, I needed to add a jest environment package.

It adds some styles to get a basic idea about the DOM.

For the other modules, it just initializes them. Most of the content
in them will need to be fixed.