Fix event listeners
Before, every time the game was played, a new set of event listeners
were added to the DOM. For some reason, removeEventListener did
not work.

To solve this, I used an AbortController to send a kill signal to the
event listener which is a bit more clear than adding the "once"
keyword to many event listeners in the code.