Fix square coordinates check and error message
The program is pretty robust in determining the straight orientation
that is required when two non-colinear points are selected. One case
that does not work is for a square. Before, I just called the user out
the user for any square, but if it is a one-by-one square, that could
be a one unit length ship.

To fix this, I made a change to the PointsHelper method to catch this
case and updated the error message.