Change water image to gif
I did not like the simple blue color for the water, so I found a picture
of water that I liked and applied a ripple animation to it using GIMP.

I was running into cascade problems, so I also needed to change the
original id>type to a class type selector so that the "miss" selector
could capture it. This involved retooling the createGrid method in the
DOM module.

I also added the attribution in the Readme.